refactor: split oversized sse.rs and vertex_ai/client.rs into submodules (#243)
sse.rs (1189 lines) -> sse/ directory:
- mod.rs: core types, parser, stream wrapper, tests (446 lines)
- openai.rs: OpenAICompatibleTransformer (123 lines)
- anthropic.rs: AnthropicTransformer (180 lines)
- gemini.rs: GeminiTransformer (185 lines)
- cohere.rs: CohereTransformer (210 lines)
- databricks.rs: DatabricksTransformer (120 lines)

vertex_ai/client.rs (1461 lines) -> client/ directory:
- mod.rs: VertexAIErrorMapper, VertexAIProvider struct, core methods (230 lines)
- operations.rs: chat completion, embedding, token counting (174 lines)
- llm_provider.rs: LLMProvider trait implementation (515 lines)
- tests.rs: all unit tests (570 lines)

Public API unchanged via mod + pub use re-exports.
