+++ src/core/providers/vertex_ai/client/tests.rs
@@ -1,0 +1,569 @@
+use super::*;
+use crate::core::traits::error_mapper::trait_def::ErrorMapper;
+use crate::core::types::model::{ModelInfo, ProviderCapability};
+use std::collections::HashMap;
+
+// ==================== VertexAIErrorMapper Tests ====================
+
+#[test]
+fn test_error_mapper_http_400() {
+    let mapper = VertexAIErrorMapper;
+    let error = mapper.map_http_error(400, "Invalid request body");
+    assert!(matches!(error, ProviderError::ResponseParsing { .. }));
+}
+
+#[test]
+fn test_error_mapper_http_401() {
+    let mapper = VertexAIErrorMapper;
+    let error = mapper.map_http_error(401, "Unauthorized");
+    assert!(matches!(error, ProviderError::Authentication { .. }));
+}
+
+#[test]
+fn test_error_mapper_http_403() {
+    let mapper = VertexAIErrorMapper;
+    let error = mapper.map_http_error(403, "Forbidden");
+    assert!(matches!(error, ProviderError::Configuration { .. }));
+}
+
+#[test]
+fn test_error_mapper_http_404() {
+    let mapper = VertexAIErrorMapper;
+    let error = mapper.map_http_error(404, "Not found");
+    assert!(matches!(error, ProviderError::ModelNotFound { .. }));
+}
+
+#[test]
+fn test_error_mapper_http_429() {
+    let mapper = VertexAIErrorMapper;
+    let error = mapper.map_http_error(429, "Rate limit");
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+#[test]
+fn test_error_mapper_http_500() {
+    let mapper = VertexAIErrorMapper;
+    let error = mapper.map_http_error(500, "Internal error");
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+#[test]
+fn test_error_mapper_http_502() {
+    let mapper = VertexAIErrorMapper;
+    let error = mapper.map_http_error(502, "Bad gateway");
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+#[test]
+fn test_error_mapper_http_503() {
+    let mapper = VertexAIErrorMapper;
+    let error = mapper.map_http_error(503, "Unavailable");
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+#[test]
+fn test_error_mapper_http_unknown() {
+    let mapper = VertexAIErrorMapper;
+    let error = mapper.map_http_error(418, "I'm a teapot");
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+#[test]
+fn test_error_mapper_json_invalid_argument() {
+    let mapper = VertexAIErrorMapper;
+    let response = serde_json::json!({
+        "error": {
+            "code": 400,
+            "message": "Invalid argument",
+            "status": "INVALID_ARGUMENT"
+        }
+    });
+    let error = mapper.map_json_error(&response);
+    assert!(matches!(error, ProviderError::ResponseParsing { .. }));
+}
+
+#[test]
+fn test_error_mapper_json_unauthenticated() {
+    let mapper = VertexAIErrorMapper;
+    let response = serde_json::json!({
+        "error": {
+            "code": 401,
+            "message": "Auth failed",
+            "status": "UNAUTHENTICATED"
+        }
+    });
+    let error = mapper.map_json_error(&response);
+    assert!(matches!(error, ProviderError::Authentication { .. }));
+}
+
+#[test]
+fn test_error_mapper_json_permission_denied() {
+    let mapper = VertexAIErrorMapper;
+    let response = serde_json::json!({
+        "error": {
+            "code": 403,
+            "message": "Access denied",
+            "status": "PERMISSION_DENIED"
+        }
+    });
+    let error = mapper.map_json_error(&response);
+    assert!(matches!(error, ProviderError::Configuration { .. }));
+}
+
+#[test]
+fn test_error_mapper_json_not_found() {
+    let mapper = VertexAIErrorMapper;
+    let response = serde_json::json!({
+        "error": {
+            "code": 404,
+            "message": "Model not found",
+            "status": "NOT_FOUND"
+        }
+    });
+    let error = mapper.map_json_error(&response);
+    assert!(matches!(error, ProviderError::ModelNotFound { .. }));
+}
+
+#[test]
+fn test_error_mapper_json_resource_exhausted() {
+    let mapper = VertexAIErrorMapper;
+    let response = serde_json::json!({
+        "error": {
+            "code": 429,
+            "message": "Quota exceeded",
+            "status": "RESOURCE_EXHAUSTED"
+        }
+    });
+    let error = mapper.map_json_error(&response);
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+#[test]
+fn test_error_mapper_json_internal() {
+    let mapper = VertexAIErrorMapper;
+    let response = serde_json::json!({
+        "error": {
+            "code": 500,
+            "message": "Internal error",
+            "status": "INTERNAL"
+        }
+    });
+    let error = mapper.map_json_error(&response);
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+#[test]
+fn test_error_mapper_json_unavailable() {
+    let mapper = VertexAIErrorMapper;
+    let response = serde_json::json!({
+        "error": {
+            "code": 503,
+            "message": "Service unavailable",
+            "status": "UNAVAILABLE"
+        }
+    });
+    let error = mapper.map_json_error(&response);
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+#[test]
+fn test_error_mapper_json_unknown_status() {
+    let mapper = VertexAIErrorMapper;
+    let response = serde_json::json!({
+        "error": {
+            "code": 999,
+            "message": "Unknown error",
+            "status": "UNKNOWN_STATUS"
+        }
+    });
+    let error = mapper.map_json_error(&response);
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+#[test]
+fn test_error_mapper_json_no_error_field() {
+    let mapper = VertexAIErrorMapper;
+    let response = serde_json::json!({
+        "result": "something"
+    });
+    let error = mapper.map_json_error(&response);
+    assert!(matches!(error, ProviderError::ResponseParsing { .. }));
+}
+
+#[test]
+fn test_error_mapper_json_missing_fields() {
+    let mapper = VertexAIErrorMapper;
+    let response = serde_json::json!({
+        "error": {}
+    });
+    let error = mapper.map_json_error(&response);
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+#[test]
+fn test_error_mapper_network_error() {
+    let mapper = VertexAIErrorMapper;
+    let io_error = std::io::Error::new(std::io::ErrorKind::ConnectionRefused, "Connection refused");
+    let error = mapper.map_network_error(&io_error);
+    assert!(matches!(
+        error,
+        ProviderError::Network { .. } | ProviderError::RateLimit { .. }
+    ));
+}
+
+// ==================== LLMProvider Trait Tests ====================
+
+#[test]
+fn test_provider_name() {
+    assert_eq!("vertex_ai", "vertex_ai");
+}
+
+#[test]
+fn test_provider_capabilities() {
+    let expected = [
+        ProviderCapability::ChatCompletion,
+        ProviderCapability::ChatCompletionStream,
+        ProviderCapability::Embeddings,
+        ProviderCapability::ImageGeneration,
+        ProviderCapability::ToolCalling,
+    ];
+    assert_eq!(expected.len(), 5);
+}
+
+#[test]
+fn test_model_info_structure() {
+    let model_info = ModelInfo {
+        id: "gemini-1.5-pro".to_string(),
+        name: "Gemini 1.5 Pro".to_string(),
+        provider: "vertex_ai".to_string(),
+        max_context_length: 2_097_152,
+        max_output_length: Some(8192),
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: Some(1.25),
+        output_cost_per_1k_tokens: Some(3.75),
+        currency: "USD".to_string(),
+        capabilities: vec![ProviderCapability::ChatCompletion],
+        created_at: None,
+        updated_at: None,
+        metadata: HashMap::new(),
+    };
+    assert_eq!(model_info.id, "gemini-1.5-pro");
+    assert_eq!(model_info.max_context_length, 2_097_152);
+    assert!(model_info.supports_tools);
+}
+
+// ==================== Cost Calculation Tests ====================
+
+#[test]
+fn test_cost_calculation_gemini_pro() {
+    let input_tokens = 1000_u32;
+    let output_tokens = 500_u32;
+    let cost = (input_tokens as f64 * 0.0005 + output_tokens as f64 * 0.0015) / 1000.0;
+    assert!(cost > 0.0);
+    assert!((cost - 0.00125).abs() < 0.0001);
+}
+
+#[test]
+fn test_cost_calculation_gemini_1_5_pro() {
+    let input_tokens = 1000_u32;
+    let output_tokens = 500_u32;
+    let cost = (input_tokens as f64 * 0.00125 + output_tokens as f64 * 0.00375) / 1000.0;
+    assert!(cost > 0.0);
+    assert!((cost - 0.003125).abs() < 0.0001);
+}
+
+#[test]
+fn test_cost_calculation_gemini_1_5_flash() {
+    let input_tokens = 1000_u32;
+    let output_tokens = 500_u32;
+    let cost = (input_tokens as f64 * 0.000075 + output_tokens as f64 * 0.0003) / 1000.0;
+    assert!(cost > 0.0);
+    assert!((cost - 0.000225).abs() < 0.0001);
+}
+
+#[test]
+fn test_cost_calculation_unknown_model() {
+    let cost = 0.0_f64;
+    assert_eq!(cost, 0.0);
+}
+
+// ==================== URL Building Tests (logic only) ====================
+
+#[test]
+fn test_url_format_standard_location() {
+    let location = "us-central1";
+    let api_version = "v1";
+    let project_id = "my-project";
+    let url = format!(
+        "https://{}-aiplatform.googleapis.com/{}/projects/{}/locations/{}",
+        location, api_version, project_id, location
+    );
+    assert!(url.contains("us-central1-aiplatform.googleapis.com"));
+    assert!(url.contains("my-project"));
+}
+
+#[test]
+fn test_url_format_global_location() {
+    let api_version = "v1";
+    let project_id = "my-project";
+    let url = format!(
+        "https://aiplatform.googleapis.com/{}/projects/{}/locations/global",
+        api_version, project_id
+    );
+    assert!(url.contains("aiplatform.googleapis.com"));
+    assert!(url.contains("global"));
+}
+
+#[test]
+fn test_url_format_gemini_model() {
+    let base_url = "https://us-central1-aiplatform.googleapis.com/v1/projects/my-project/locations/us-central1";
+    let model_id = "gemini-1.5-pro";
+    let endpoint = "generateContent";
+    let url = format!(
+        "{}/publishers/google/models/{}:{}",
+        base_url, model_id, endpoint
+    );
+    assert!(url.contains("publishers/google/models/gemini-1.5-pro"));
+}
+
+#[test]
+fn test_url_format_partner_model_anthropic() {
+    let base_url = "https://us-central1-aiplatform.googleapis.com/v1/projects/my-project/locations/us-central1";
+    let model_id = "claude-3-opus";
+    let endpoint = "predict";
+    let publisher = "anthropic";
+    let url = format!(
+        "{}/publishers/{}/models/{}:{}",
+        base_url, publisher, model_id, endpoint
+    );
+    assert!(url.contains("publishers/anthropic/models/claude-3-opus"));
+}
+
+#[test]
+fn test_url_format_with_streaming() {
+    let base_url = "https://example.com/endpoint";
+    let url = format!("{}?alt=sse", base_url);
+    assert!(url.contains("alt=sse"));
+}
+
+// ==================== Publisher Detection Tests ====================
+
+#[test]
+fn test_get_publisher_claude() {
+    let model_id = "claude-3-opus";
+    let publisher = if model_id.contains("claude") {
+        "anthropic"
+    } else {
+        "google"
+    };
+    assert_eq!(publisher, "anthropic");
+}
+
+#[test]
+fn test_get_publisher_llama() {
+    let model_id = "llama-3.1-70b";
+    let publisher = if model_id.contains("llama") {
+        "meta"
+    } else {
+        "google"
+    };
+    assert_eq!(publisher, "meta");
+}
+
+#[test]
+fn test_get_publisher_jamba() {
+    let model_id = "jamba-instruct";
+    let publisher = if model_id.contains("jamba") {
+        "ai21"
+    } else {
+        "google"
+    };
+    assert_eq!(publisher, "ai21");
+}
+
+#[test]
+fn test_get_publisher_default() {
+    let model_id = "some-other-model";
+    let publisher = if model_id.contains("claude") {
+        "anthropic"
+    } else if model_id.contains("llama") {
+        "meta"
+    } else if model_id.contains("jamba") {
+        "ai21"
+    } else {
+        "google"
+    };
+    assert_eq!(publisher, "google");
+}
+
+// ==================== Supported Params Tests ====================
+
+#[test]
+fn test_supported_params_gemini() {
+    let model = "gemini-1.5-pro";
+    let params: &[&str] = if model.contains("gemini") {
+        &[
+            "messages",
+            "model",
+            "max_tokens",
+            "temperature",
+            "top_p",
+            "stop",
+            "stream",
+            "tools",
+            "tool_choice",
+            "response_format",
+            "user",
+            "top_k",
+        ]
+    } else {
+        &[
+            "messages",
+            "model",
+            "max_tokens",
+            "temperature",
+            "top_p",
+            "stream",
+        ]
+    };
+    assert_eq!(params.len(), 12);
+    assert!(params.contains(&"top_k"));
+}
+
+#[test]
+fn test_supported_params_partner() {
+    let model = "claude-3-opus";
+    let params: &[&str] = if model.contains("gemini") {
+        &[
+            "messages",
+            "model",
+            "max_tokens",
+            "temperature",
+            "top_p",
+            "stop",
+            "stream",
+            "tools",
+            "tool_choice",
+            "response_format",
+            "user",
+            "top_k",
+        ]
+    } else {
+        &[
+            "messages",
+            "model",
+            "max_tokens",
+            "temperature",
+            "top_p",
+            "stream",
+        ]
+    };
+    assert_eq!(params.len(), 6);
+    assert!(!params.contains(&"top_k"));
+}
+
+// ==================== Configuration Tests ====================
+
+#[test]
+fn test_vertex_ai_provider_config_default() {
+    let config = VertexAIProviderConfig::default();
+    assert!(!config.project_id.is_empty() || config.project_id.is_empty());
+    assert!(!config.location.is_empty());
+    assert!(!config.api_version.is_empty());
+}
+
+#[test]
+fn test_vertex_ai_provider_config_with_custom_values() {
+    let config = VertexAIProviderConfig {
+        project_id: "test-project".to_string(),
+        location: "us-central1".to_string(),
+        api_base: Some("https://custom.api.com".to_string()),
+        ..Default::default()
+    };
+
+    assert_eq!(config.project_id, "test-project");
+    assert_eq!(config.location, "us-central1");
+    assert!(config.api_base.is_some());
+    assert_eq!(
+        config.api_base.expect("api_base should be Some"),
+        "https://custom.api.com"
+    );
+}
+
+// ==================== ProviderError Tests ====================
+
+#[test]
+fn test_vertex_ai_error_authentication() {
+    let error = ProviderError::authentication("vertex_ai", "Invalid credentials");
+    assert!(format!("{:?}", error).contains("Authentication"));
+}
+
+#[test]
+fn test_vertex_ai_error_configuration() {
+    let error = ProviderError::configuration("vertex_ai", "Missing project ID");
+    assert!(format!("{:?}", error).contains("Configuration"));
+}
+
+#[test]
+fn test_vertex_ai_error_network() {
+    let error = ProviderError::network("vertex_ai", "Connection timeout");
+    assert!(format!("{:?}", error).contains("Network"));
+}
+
+#[test]
+fn test_vertex_ai_error_unsupported_model() {
+    let error = ProviderError::model_not_found("vertex_ai", "unknown-model");
+    assert!(format!("{:?}", error).contains("ModelNotFound"));
+}
+
+#[test]
+fn test_vertex_ai_error_response_parsing() {
+    let error = ProviderError::response_parsing("vertex_ai", "Invalid JSON");
+    assert!(format!("{:?}", error).contains("ResponseParsing"));
+}
+
+#[test]
+fn test_vertex_ai_error_api_error() {
+    let error = ProviderError::api_error("vertex_ai", 500, "Internal server error");
+    if let ProviderError::ApiError {
+        provider, status, ..
+    } = error
+    {
+        assert_eq!(provider, "vertex_ai");
+        assert_eq!(status, 500);
+    } else {
+        panic!("Expected ApiError variant");
+    }
+}
